More secure API

+++ application/classes/Controller/Api/Marker.php
@@ -6,6 +6,8 @@ class Controller_Api_Marker extends Controller_Rest {
 	{
 		$data = (array) json_decode($this->request->body());
 
+		$data['user_id'] = $this->user->id;
+
 		$marker = ORM::factory('Marker');
 		$marker->values($data);
 		$marker->save();
@@ -27,6 +29,11 @@ class Controller_Api_Marker extends Controller_Rest {
 			throw HTTP_Exception::factory(500, 'Marker not found');
 		}
 
+		if ($marker->user_id !== $this->user->id)
+		{
+			throw HTTP_Exception::factory(401);
+		}
+
 		$marker->values(array(
 			'longitude' => $data['longitude'],
 			'latitude' => $data['latitude']
@@ -45,6 +52,11 @@ class Controller_Api_Marker extends Controller_Rest {
 		if (!$marker->loaded())
 		{
 			throw HTTP_Exception::factory(500, 'Marker not found');
+		}
+
+		if ($marker->user_id !== $this->user->id)
+		{
+			throw HTTP_Exception::factory(401);
 		}
 
 		$marker->delete();
